Listo todo menos UDP, falta pedidos no satisfechos en CP

+++ p2/CL/lib/string_lib.h
@@ -1,6 +1,7 @@
 #include <iostream>
 #include <sstream>
 #include <string>
+#include <cstdlib>
 
 using namespace std;
 
@@ -9,5 +10,5 @@ const string trim(const string& pString);
 string int_to_str(int n);
 string double_to_str(double n);
 int str_to_int(string s);
-int str_to_double(string s);
+double str_to_double(string s);
 char ** split(char string[], int * num, char * sep);
